% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{boot_sample}
\alias{boot_sample}
\title{boot_sample
Generate n bootstrap sample of the monitoring sites to be used for each iteration.}
\usage{
boot_sample(data, boot_n = 1000)
}
\arguments{
\item{data}{data.table or data.frame Data with all site id.}

\item{boot_n}{integer The number of bootstrap samples to be generated.}
}
\value{
A list with site id and bootstrap indices for n bootstrap sample.
}
\description{
boot_sample
Generate n bootstrap sample of the monitoring sites to be used for each iteration.
}
\seealso{
\link{impute_count}, \link{collated_index}
}
\author{
Reto Schmucki - \email{retoshm@ceh.ac.uk}
}
\keyword{bootstrap}
\keyword{collated}
\keyword{index}
