% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{butterfly_day}
\alias{butterfly_day}
\title{butterfly_day
count cumulative butterfly count observed over one monitoring season.}
\usage{
butterfly_day(sp_ts_season_count, WeekCount = TRUE)
}
\arguments{
\item{sp_ts_season_count}{list of objects, including a data.table with observed and expected butterfly counts per day imputed for each sites based by \link{impute_count}.}

\item{WeekCount}{logical defining if butterfly day should be counted by week (default), using the 4th day of the week, or as a total of predicted daily counts.}
}
\value{
data.table with total butterfly day observed per species, year, and site.
}
\description{
butterfly_day
count cumulative butterfly count observed over one monitoring season.
}
\seealso{
\link{impute_count}, \link{flight_curve}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{butterfly}
\keyword{count}
