% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_toolbox.r
\name{check_names}
\alias{check_names}
\title{check_names
Verify for the required column names in the data object.}
\usage{
check_names(x, y)
}
\arguments{
\item{x}{Data object in which the variables names are verified.}

\item{y}{Variable names to be verified.}
}
\value{
Verifyies if column names listed in \code{y} are found in the data set \code{x}, if not, a message identifies the
missing column name and stops.
}
\description{
check_names
Verify for the required column names in the data object.
}
\details{
This function is not case sensitive, but it does not accept different names or spelling.
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
