% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_toolbox.r
\name{check_package}
\alias{check_package}
\title{check_package
Internal function to check if a package is installed.}
\usage{
check_package(
  pkgName = NULL,
  message1 = "You need to install ",
  message2 = "This version requires"
)
}
\arguments{
\item{pkgName}{The package name to be verified.}

\item{message1}{Inform the user about the package dependency.}

\item{message2}{Inform the user what happen if the package is not installed.}
}
\value{
If package is not found, the user is offered the option to install the missing package.
}
\description{
check_package
Internal function to check if a package is installed.
}
\author{
Reto Schmucki - \email{retoshm@ceh.ac.uk}
}
