% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{collated_index}
\alias{collated_index}
\title{collated_index
compute a collated index from the site indices, using a Generalized Linear Model.}
\usage{
collated_index(data, s_sp, sindex_value = "SINDEX", bootID = NULL,
  boot_ind = boot_ind, glm_weights = TRUE, rm_zero = TRUE)
}
\arguments{
\item{data}{data.table or data.frame with site indices per year and proportion of fligth curve covered by the monitoring.}

\item{s_sp}{string with with the species name to be found in the data.}

\item{sindex_value}{string defining the response variable to be used for collated index computation, default is "SINDEX", but could be other standardized values.}

\item{bootID}{integer with the n^th bootstrap.}

\item{boot_ind}{data.table with in index of the bootstrap and the corresponding set of site id to be used for the specific bootstrap sample}

\item{glm_weights}{logical if collated index should be calculated with weighting for the proportion of the flight curve, if FALSE, equal weigth is used.}

\item{rm_zero}{logical specifying if site where species has not been observed should be removed from the GLM fitting, this speed up the process without affecting the output, default TRUE.}
}
\value{
a list of two objects, a vector of site, a glm model object, and a vector of collated indices per year.
}
\description{
collated_index
compute a collated index from the site indices, using a Generalized Linear Model.
}
\seealso{
\link{impute_count}, \link{flight_curve}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{annual}
\keyword{index}
