% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{collated_index_old}
\alias{collated_index_old}
\title{collated_index_old
compute a collated index from the site indices, using a Generalized Linear Model.}
\usage{
collated_index_old(site_indices, GlmWeight = NULL, GlmFamily = poisson())
}
\arguments{
\item{site_indices}{data.table or data.frame with site indices per year and proportion of flight curve covered by the monitoring.}

\item{GlmWeight}{vector of weight used in the GLM.}

\item{GlmFamily}{family used for the GLM model.}
}
\value{
a list of three objects, a vector of site, a glm model object, and a vector of collated indices per year.
}
\description{
collated_index_old
compute a collated index from the site indices, using a Generalized Linear Model.
}
\seealso{
\link{impute_count}, \link{flight_curve}
}
\author{
Reto Schmucki - \email{retoshm@ceh.ac.uk}
}
\keyword{annual}
\keyword{index}
