% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{day_week_summary}
\alias{day_week_summary}
\title{day_week_summary
Summarize the count(s) per day or week using either the maximum or the average count when multiple counts (visit) are observed.}
\usage{
day_week_summary(ts_season_count, MultiVisit, TimeUnit, weekday)
}
\arguments{
\item{ts_season_count}{data.table Time-series of butterfly counts for species x over year y over all sites.}

\item{MultiVisit}{string Function to apply for summarising multiple counts within a time unit, 'max' or 'mean' (default).}

\item{TimeUnit}{character The time-step for which the spline should be computed, 'd' day or 'w' week. For TimeUnit week, 'w', we refer to week_day 3 to match a mid-week calendar date.}

\item{weekday}{Integer for selected day of the week for weekly summary, default is 3 (Wednesday). [1-7] where 1 = Monday.}
}
\value{
data.table Time-series of butterfly counts for species x over year y over all sites, but summarized with the chosen function.
}
\description{
day_week_summary
Summarize the count(s) per day or week using either the maximum or the average count when multiple counts (visit) are observed.
}
\seealso{
\link{flight_curve}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{flight_curve}
