% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{fit_gam}
\alias{fit_gam}
\title{fit_gam
fit a Generalized Additive Model to butterfly count data along a temporal variable and accounting for site effect when multiple are available.}
\usage{
fit_gam(
  dataset_y,
  NbrSample = NULL,
  GamFamily = "poisson",
  MaxTrial = 4,
  SpeedGam = TRUE,
  OptiGam = TRUE,
  ConLikelihood = TRUE,
  TimeUnit = "d",
  MultiVisit = "mean"
)
}
\arguments{
\item{dataset_y}{data.table with filtered butterfly counts for species x over year y over all sites.}

\item{NbrSample}{integer inherited from \link{flight_curve}, default=100.}

\item{GamFamily}{string inherited from \link{flight_curve}, default='poisson', but can be 'nb' or 'quasipoisson'.}

\item{MaxTrial}{integer inherited from \link{flight_curve}, default=3.}

\item{SpeedGam}{Logical to use the \link[mgcv]{bam} method instead of the \link[mgcv]{gam} method.}

\item{OptiGam}{Logical to set use bam when data are larger than 100 and gam for smaller dataset}

\item{TimeUnit}{Character defining if the spline should be computed at the day 'd' or the week 'w'.}

\item{MultiVisit}{function to be applied on multiple count within a single time unit, 'max' or 'mean' (default).}

\item{...}{additional parameters passed to gam or bam function from the \link[mgcv]{gam} package.}
}
\value{
A list with three objects, i) **f_curve**: a data.table with the flight curve \code{f_curve} with expected relative abundance, normalize to sum to one over a full season,
        ii) **f_model**: the resulting gam model \code{f_model} fitted on the count data and iii) **f_data**: a data.table with the data used to fit the GAM model. This is provide for one year 'y'.
}
\description{
fit_gam
fit a Generalized Additive Model to butterfly count data along a temporal variable and accounting for site effect when multiple are available.
}
\seealso{
\link{flight_curve}, \link[mgcv]{gam}, \link[mgcv]{bam}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{gam}
