% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{fit_speedglm}
\alias{fit_speedglm}
\title{fit_speedglm
Fit a Generalized Linear Model (GLM) to predict daily count per site, using the regional flight curve as offset and the negative binomial error distribution 
as implemented in the package \code{link[speedglm{speedglm}}. Function used in \code{\link{impute_count}}.}
\usage{
fit_speedglm(sp_count_flight_y, non_zero, FamilyGlm)
}
\arguments{
\item{sp_count_flight_y}{data.table with time series of the expected relative abundance of butterfly count per day (NM) for the year of interest,
or the nearest available.}

\item{non_zero}{vector of sites with non-zero value to be included in the model (see detail).}

\item{FamilyGlm}{string for the distribution to be used for the error term in the GLM, inherited from \code{\link{impute_count}}, default='quasipoisson'.}
}
\value{
A a list of object, i) data.table with data and expected relative abundance of butterfly count per day (NM) for the year or the nearest year where 
        a fligh curve (phenology) is available \code{sp_count_flight_y} and ii) a glm object for the model \code{glm_obj_site}.
}
\description{
fit_speedglm
Fit a Generalized Linear Model (GLM) to predict daily count per site, using the regional flight curve as offset and the negative binomial error distribution 
as implemented in the package \code{link[speedglm{speedglm}}. Function used in \code{\link{impute_count}}.
}
\details{
GLM model is only fitted for site with observations, non-zero, as the abundance for sites where count are only zeros is expected to be null
}
\seealso{
\code{\link{impute_count}}, \code{\link{flight_curve}}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{curve}
\keyword{flight}
