% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_toolbox.r
\name{get_bioclim}
\alias{get_bioclim}
\title{get_bioclim
Assign a bioclimatic region to the value extractacted from raster layer for a set of geographic points}
\usage{
get_bioclim(x, y = "metzger_v3_class", byY = "gens_seq", xCrs = 4326,
  OutDf = TRUE)
}
\arguments{
\item{x}{data.frame, or sf object, with longitude and latitude in column 1 and 2 respectively and bioclim value extracted with get_raster_value}

\item{y}{data.frame with bioclimatic region classification key refering to the raster layer used by get_raster_value}

\item{byY}{character name of the variable name corresponding to the raster layer value in y}

\item{xCrs}{EPSG number of the cordinate reference system (CRS) of the x coordinates, see \code{\url{http://spatialreference.org}}}

\item{OutDf}{logical if true return a data.frame, if false, a sf object is returned}
}
\value{
return a data.frame or a sf object with bioclimatic region extracted from the y layer.
}
\description{
get_bioclim
Assign a bioclimatic region to the value extractacted from raster layer for a set of geographic points
}
\examples{
x <- data.frame(longitude = c(4, 4.1, 4.5), latitude = c(50, 50.45, 50.5), id = c('a','b','c'))
x_value <- get_raster_value(x, OutDf = FALSE)
get_bioclim(x_value)
get_bioclim(x_value, OutDf = FALSE)

}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
