% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{get_nny}
\alias{get_nny}
\title{get_nny
find the nearest year with a computed flight curve}
\usage{
get_nny(x, y)
}
\arguments{
\item{x}{integer Year for which to find the nearest flight curve.}

\item{y}{vector Years with available flight curve.}
}
\value{
z integer The value of the nearest year with flight curve.
}
\description{
get_nny
find the nearest year with a computed flight curve
}
\seealso{
\link{check_pheno}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{curve}
\keyword{flight}
