% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_toolbox.r
\name{get_rater_value}
\alias{get_rater_value}
\title{get_rater_value 
Extract raster value for a set of geographic points}
\usage{
get_rater_value(x, y = metzger_v3_europe, Classification = NULL,
  x_crs = 4326, buffer_dist = NULL, out_df = TRUE)
}
\arguments{
\item{x}{data frame with longitude and latitude in column 1 and 2 respectively}

\item{y}{raster layer from which value should be extracted}

\item{Classification}{a data.frame with classification for the raster values}

\item{x_crs}{EPSG number of the cordinate reference system (CRS) of the x coordinates, see \code{\link{http://spatialreference.org/}}}

\item{buffer_dist}{extent of the buffer to be added around the bounding box to insure}

\item{out_df}{logical if true return a data.frame, if false, a sf object is returned}
}
\value{
return a df or a sf object with value extracted from the y layer, see \code{out_df}.
}
\description{
get_rater_value 
Extract raster value for a set of geographic points
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
