% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{impute_count}
\alias{impute_count}
\title{impute_count}
\usage{
impute_count(
  ts_season_count,
  ts_flight_curve,
  TimeUnit,
  sp = NULL,
  YearLimit = NULL,
  SelectYear = NULL,
  CompltSeason = TRUE
)
}
\arguments{
\item{ts_season_count}{data.table with time series of counts for a specific species across all sites as returned by \link{ts_monit_count_site}.}

\item{ts_flight_curve}{data.table with the flight curves, relative abundance (NM), for a specific species as returned by \link{flight_curve}.}

\item{TimeUnit}{Character to define days 'd' or week 'w' as variable for the GAM.}

\item{sp}{integer or string for the species ID or name.}

\item{YearLimit}{integer defining the range (+/- number of year) of year to look for a flight curve, if NULL no restriction is set.}

\item{SelectYear}{integer to select a specific year to compute the flight curve, default=NULL.}

\item{CompltSeason}{Logical to restrict computation of flight curve for years where the complete season has been sampled, default=TRUE.}
}
\value{
A data.table based on the entry count data, augmented with site indices 'SINDEX' and imputed weekly count 'IMPUTED_COUNT'.
}
\description{
impute_count
}
\details{
Site indices can be extracted from the data.table returned from this function. The Site index is currently computed by adjusting the count by the proportion of the flight curve covered by the visits.
}
\seealso{
\link{flight_curve}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{curve}
\keyword{flight}
\keyword{index,}
\keyword{site}
