% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_toolbox.r
\name{initiate_project}
\alias{initiate_project}
\title{initiate_project
Build the folder structure for a generic research project}
\usage{
initiate_project(project_name, project_home = NULL)
}
\arguments{
\item{project_name}{Name of the project to be used for the parent folder}

\item{project_home}{Path where the project folder structure to be built, default is the current working directory}
}
\description{
initiate_project
Build the folder structure for a generic research project
}
\author{
Reto Schmucki - \email{retoshm@ceh.ac.uk}
}
