% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{m_count}
\alias{m_count}
\title{Toy data set with butterfly count for x species across y sites}
\format{
data.table object with butterfly count in long format
\describe{
  \item{SITE_ID}{number used as site id, the can be and integer or a string}
  \item{DATE}{date when the observation/count was recorded, format YEAR-MM-DD}
  \item{SPECIES}{number used as species id, this can be an integer or a string (e.g. "Aglais io")}
  \item{DAY}{integer, day within the month}
  \item{MONTH}{integer, month within the year}
  \item{YEAR}{integer, 4 digit year}
  \item{COUNT}{integer, actual butterfly count}
}
}
\usage{
m_count
}
\description{
Toy data set with butterfly count for x species across y sites
}
\keyword{datasets}
