% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{metzger_v3_class}
\alias{metzger_v3_class}
\title{Classification of climatic bioclimatic regions used in Metzger's high-resolution bioclimate map of the world.}
\format{Integer associated classification:
\describe{
  \item{value}{Integer associated classification found in the \code{metzger_v3_europe_values} dataset}
}}
\source{
Global map is available from \url{https://edinburgh-innovations.ed.ac.uk/project/bioclimate-world-map}.
}
\usage{
metzger_v3_class
}
\description{
Classification of climatic bioclimatic regions used in Metzger's high-resolution bioclimate map of the world.
}
\references{
Metzger, M.J., Bunce, R.G.H., Jongman, R.H.G., Sayre, R., Trabucco, A. & Zomer, R. (2013) A high-resolution bioclimate map of the world: a unifying framework
for global biodiversity research and monitoring. Global Ecology and Biogeography, 22, 630–638.
}
\keyword{datasets}
