% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_methods.r
\name{plot.pheno_curve}
\alias{plot.pheno_curve}
\title{plot.pheno_curve
Generic method to plot the flight curve, where values are extracted from a pheno_curve object (outcome of the light_curve() function)}
\usage{
\method{plot}{pheno_curve}(data, year = NULL, weekday = 3, ...)
}
\arguments{
\item{data}{pheno_curve object which is the outcome of the light_curve() function}

\item{year}{integer for the year to be displayed (e.g. 2015), default is NULL.}

\item{weekday}{weekday to be used for date in weekly count, where 1 refers to Monday, default is 3 (Wednesday)}

\item{...}{additional parameters for base plot.}
}
\value{
Returns a base plot with relative abundance (y) over time (x)
}
\description{
plot.pheno_curve
Generic method to plot the flight curve, where values are extracted from a pheno_curve object (outcome of the light_curve() function)
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
