% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_organize_data.r
\name{set_anchor}
\alias{set_anchor}
\title{set_anchor
Add Anchors of "zeros" at determined distance on each side of the monitoring season with specific weight (length),
this function is used by ts_monit_season()}
\usage{
set_anchor(FirstObs, LastObs, AnchorLength = 7, AnchorLag = 7,
  TimeUnit = "d")
}
\arguments{
\item{FirstObs}{integer defining the start of the monitoring season - correspond to the day since}

\item{LastObs}{integer defining the end of the monitoring season - correspond to the day since}

\item{AnchorLength}{integer defining the number of days used as Anchor each side of the monitoring season}

\item{AnchorLag}{integer defining the number of days between the Anchor and the monitoring season}

\item{TimeUnit}{character defining the time unit used for the Anchor, can be 'd' for day or 'w' for week}
}
\description{
set_anchor
Add Anchors of "zeros" at determined distance on each side of the monitoring season with specific weight (length),
this function is used by ts_monit_season()
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
