% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_toolbox.r
\name{ts_date_seq}
\alias{ts_date_seq}
\title{ts_date_seq
Generate a time-series of dates (per day) from the beginning of a starting year to the end of an ending years.}
\usage{
ts_date_seq(InitYear = 1970, LastYear = format(Sys.Date(), "\%Y"))
}
\arguments{
\item{InitYear}{start year of the time-series, 4 numbers format (e.g. 1987)}

\item{LastYear}{end year of the time-series, if not provided, current year is used instead}
}
\value{
return a POSIXct vector with the format 'YYYY-MM-DD HH:MM:SS'
}
\description{
ts_date_seq
Generate a time-series of dates (per day) from the beginning of a starting year to the end of an ending years.
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{series}
\keyword{time}
