% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_organize_data.r
\name{ts_dwmy_table}
\alias{ts_dwmy_table}
\title{ts_dwmy_table
Generate a time-series of dates with day, week, month and year (dwmy) from one initial to an end years}
\usage{
ts_dwmy_table(InitYear = 1970, LastYear = format(Sys.Date(), "\%Y"),
  WeekDay1 = "monday")
}
\arguments{
\item{InitYear}{start year of the time-series, 4 numbers format (e.g 1987)}

\item{LastYear}{end year of the time-series, if not provided, current year is used instead}

\item{WeekDay1}{to start the week on monday, use 'monday', otherwise the week start on sunday}
}
\value{
a data.table object with the date, the day since the first date, the year, the month,
         the day in the month, the ISO week number and the day in the week.
}
\description{
ts_dwmy_table
Generate a time-series of dates with day, week, month and year (dwmy) from one initial to an end years
}
\seealso{
\link{ts_date_seq}, \link[data.table]{IDate}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{series}
\keyword{time}
