% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_organize_data.r
\name{ts_monit_count_site}
\alias{ts_monit_count_site}
\title{ts_monit_count_site
Generate a full time series of observed count, for all sites and each day since a starting and ending years of the defined time-series}
\usage{
ts_monit_count_site(m_season_visit, m_count, sp = 1,
  DateFormat = "\%Y-\%m-\%d")
}
\arguments{
\item{m_season_visit}{data.table with potential absence based on visit data as returned by \code{ts_monit_site}}

\item{m_count}{data.table or data.frame with monitoring count data}

\item{sp}{integer or string for the species ID or name}

\item{DateFormat}{format used for the date in the count data, default="\%Y-\%m-\%d".}
}
\value{
A data.table with a complete time-series where absences (0) and recorded counts for a specific species are implemented for each 
        site.
}
\description{
ts_monit_count_site
Generate a full time series of observed count, for all sites and each day since a starting and ending years of the defined time-series
}
\seealso{
\code{\link{ts_monit_site}}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
