% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_organize_data.r
\name{ts_monit_season}
\alias{ts_monit_season}
\title{ts_monit_season
Build a time-series of dates with specific detail about the monitoring season}
\usage{
ts_monit_season(d_series, StartMonth = 4, EndMonth = 9, StartDay = 1,
  EndDay = NULL, CompltSeason = TRUE, Anchor = TRUE, AnchorLength = 7,
  AnchorLag = 7)
}
\arguments{
\item{d_series}{time series of dates returned by \code{\link{ts_dwmy_table}}}

\item{StartMonth}{integer for the first month of the monitoring season, default=4}

\item{EndMonth}{integer for the last month of the monitoring season, default=9}

\item{StartDay}{integer for the day of the month when the monitoring season start, default=1}

\item{EndDay}{integer for the day of the month when the monitoring season end, default=last day of the EndMonth}

\item{CompltSeason}{logical if only the date for a complete season should be returned, default=TRUE}

\item{Anchor}{logical if Anchor should be used at the beginning and end of the monitoring season, default=TRUE}

\item{AnchorLength}{integer for the number of day used as Anchor, default=7}

\item{AnchorLag}{integer for the number of days before and after where the Anchor should start, default=TRUE}
}
\value{
A data.table with the entire time-series of date (\code{DATE}, \code{DAY_SINCE}, \code{YEAR}, \code{MONTH},
        \code{DAY}, \code{WEEK} (ISO), the \code{WEEK_DAY}, and details about the Monitoring Year (\code{M_YEAR}) which
        is the monitoring year to which that date refers to, using the year of the starting month of the monitoring season,
        the monitoring season (\code{M_SEASON}), if the Monitoring season is complete (\code{COMPLT_SEASON}), the
        location of the anchors (\code{ANCHOR}) and the count for every day used as anchor (0).
}
\description{
ts_monit_season
Build a time-series of dates with specific detail about the monitoring season
}
\details{
The monitoring season can start in year y and end in year y+1, so if the monitoring season goes over two 
        years (i.e. winter, December-January) the SEASON YEAR is shifted to set the monitoring season within a 
        continuous year named according to the year at the start. Here the monitoring season is set in the middle 
        or year' to give some room to set ANCHORs  before and after the monitoring season.
}
\seealso{
\code{\link{ts_date_seq}}, \code{\link{ts_dwmy_table}}, \code{\link{set_anchor}}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
