% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_organize_data.r
\name{ts_monit_site}
\alias{ts_monit_site}
\title{ts_monit_site
Augment the time series in m_season with all sites and visits with "zeros", leaving all non visited day with and <NA>}
\usage{
ts_monit_site(m_visit, ts_season, DateFormat = "\%Y-\%m-\%d")
}
\arguments{
\item{m_visit}{data.table or data.frame with Date and Site ID for each monitoring visit.}

\item{ts_season}{data.table returned by \link{ts_monit_season} with the detail time-series of the monitoring season.}

\item{DateFormat}{format used for the date in the visit data, default="\%Y-\%m-\%d".}
}
\value{
A data.table with a complete time-series where absences have been implemented and that is ready to receive counts
}
\description{
ts_monit_site
Augment the time series in m_season with all sites and visits with "zeros", leaving all non visited day with and <NA>
}
\seealso{
\link{df_visit_season}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
