% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{boot_sample}
\alias{boot_sample}
\title{boot_sample
generat n bootstrap sample from the site indices.}
\usage{
boot_sample(data, boot_n = 1000)
}
\arguments{
\item{data}{data.table or data.frame with site id.}

\item{boot_n}{integer defining the number of bootstrap sample to generate.}
}
\value{
a list with site id and bootstrap indices for n bootstrap sample.
}
\description{
boot_sample
generat n bootstrap sample from the site indices.
}
\seealso{
\link{impute_count}, \link{collated_index}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{bootstrap}
\keyword{collated}
\keyword{index}
