% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{butterfly_day}
\alias{butterfly_day}
\title{butterfly_day
count cumulative butterfly count observed over one monitoring season, including observed and imputed values.}
\usage{
butterfly_day(sp_ts_season_count)
}
\arguments{
\item{sp_ts_season_count}{data.table with observed and expected butterfly counts per day imputed for each sites based by \code{\link{impute_count}}.}
}
\value{
data.table with total butterfly day observed per species, year, and site.
}
\description{
butterfly_day
count cumulative butterfly count observed over one monitoring season, including observed and imputed values.
}
\seealso{
\code{\link{impute_count}}, \code{\link{flight_curve}}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{butterfly}
\keyword{count}
