% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_toolbox.r
\name{check_names}
\alias{check_names}
\title{check_names
Verify for the required column names in the data}
\usage{
check_names(x, y)
}
\arguments{
\item{x}{a data.table object with column names}

\item{y}{vector with the required variable names}
}
\value{
Verify if column names listed in \code{y} vector are found in the data set \code{x}, if not, a message identifies the 
missing column name and stops.
}
\description{
check_names
Verify for the required column names in the data
}
\details{
This function is not case sensitive, but it does not accept different names or spelling.
}
\examples{
DF <- data.frame(DAY=c(1:5),MONTH=rep(month.name[2],5),YEAR=rep(format(Sys.Date(),'\%Y'),5))
check_names(DT,c('DAY','month','Years'))
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
