% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_toolbox.r
\name{check_package}
\alias{check_package}
\title{check_package
Internal function to verified the required package is installed}
\usage{
check_package(
  pkgName = NULL,
  message1 = "you need to install the package ",
  message2 = "This version requires "
)
}
\arguments{
\item{pkgName}{A string with the package name}

\item{message1}{A string to inform about the dependency}

\item{message2}{A string to inform what happen if not installed}
}
\value{
If package is not installed, the function ask to install the package.
}
\description{
check_package
Internal function to verified the required package is installed
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
