% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{check_pheno}
\alias{check_pheno}
\title{check_pheno
Check for the flight curve of a specific year. If the specific year is missing, use the nearest year available within a 5-year period to impute missing count. Function used in \link{impute_count}.}
\usage{
check_pheno(sp_count_flight, ts_flight_curve, YearCheck, YearLimit, TimeUnit)
}
\arguments{
\item{sp_count_flight}{data.table Object with the flight curve, relative abundance (NM), for a specific year.}

\item{ts_flight_curve}{data.table Object with the all flight curves and relative abundance (NM) for the years available for the search as returned by \link{flight_curve}.}

\item{YearCheck}{integer or vector Year to check for nearest flight curve, set internally in \link{impute_count}.}

\item{YearLimit}{integer Define the span of years (+/- number of year) to look for a flight curve, if NULL no restriction is set.}

\item{TimeUnit}{character The time-step for which the spline should be computed, 'd' day or 'w' week.}
}
\value{
A data.table with time-series of the expected relative abundance of butterfly count per day (NM) for the year or the nearest year where
        phenology is available.
}
\description{
check_pheno
Check for the flight curve of a specific year. If the specific year is missing, use the nearest year available within a 5-year period to impute missing count. Function used in \link{impute_count}.
}
\seealso{
\link{impute_count}, \link{flight_curve}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{curve}
\keyword{flight}
