% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{collated_index}
\alias{collated_index}
\title{collated_index
compute a collated index from the site indices, using a Generalized Linear Model.}
\usage{
collated_index(
  data,
  s_sp,
  sindex_value = "SINDEX",
  bootID = NULL,
  boot_ind = NULL,
  glm_weights = TRUE,
  rm_zero = TRUE
)
}
\arguments{
\item{data}{data.table or data.frame Site indices per year and proportion of flight curve covered by the monitoring.}

\item{s_sp}{string Species name to be found in the data.}

\item{sindex_value}{string Name of the response variable to be used the computation of the collated index, default is "SINDEX", but could be other standardized values.}

\item{bootID}{integer Identify the n^th bootstrap.}

\item{boot_ind}{data.table Index of the bootstrap and the site ids of the specific bootstrap sample.}

\item{glm_weights}{logical Use the proportion of the flight curve sampled as weight for the collated index, if FALSE the function uses equal weights.}

\item{rm_zero}{logical Remove the sites where species was not observed to speed-up the fit of the GLM without altering the output, default TRUE.}
}
\value{
a list of two objects, a vector of site, a glm model object, and a vector of collated indices per year.
}
\description{
collated_index
compute a collated index from the site indices, using a Generalized Linear Model.
}
\seealso{
\link{impute_count}, \link{flight_curve}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{annual}
\keyword{index}
