% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{fit_gam}
\alias{fit_gam}
\title{fit_gam
fit a Generalized Additive Model to butterfly count data to extract the annual flight curve.}
\usage{
fit_gam(dataset_y, NbrSample = NbrSample, GamFamily = GamFamily,
  MaxTrial = MaxTrial, SpeedGam = TRUE, OptiGam = TRUE, ...)
}
\arguments{
\item{dataset_y}{data.table with filtered butterfly counts for species x over year y over all sites.}

\item{NbrSample}{integer inherited from \code{link{flight_curve}}, default=100.}

\item{GamFamily}{string inherited from \code{link{flight_curve}}, default='poisson', but can be 'nb' or 'quasipoisson'.}

\item{MaxTrial}{integer inherited from \code{link{flight_curve}}, default=3.}

\item{SpeedGam}{Logical to use the \code{link[mgcv]{bam}} method instead of the \code{link[mgcv]{gam}} method.}

\item{OptiGam}{Logical to set use bam when data are larger than 100 and gam for smaller dataset}

\item{...}{additional parameters passed to gam or bam function from the \code{link[mgcv]{gam}} package.}
}
\value{
A list with two object, i) a data.table with the fligth curve \code{f_curve} with expected relative abudance, normalize to sum to one over a full season,
        and ii) the resulting gam model \code{f_model} fitted on the count data.
}
\description{
fit_gam
fit a Generalized Additive Model to butterfly count data to extract the annual flight curve.
}
\seealso{
\code{\link{flight_curve}}, \code{link[mgcv]{gam}}, \code{link[mgcv]{bam}}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{gam}
