% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{fit_gam}
\alias{fit_gam}
\title{fit_gam
fit a Generalized Additive Model to butterfly count data along a temporal variable and accounting for site effect when multiple are available.}
\usage{
fit_gam(
  dataset_y,
  NbrSample = NULL,
  GamFamily = "poisson",
  MaxTrial = 4,
  SpeedGam = TRUE,
  OptiGam = TRUE,
  TimeUnit = "d",
  MultiVisit = "mean",
  weekday = 3,
  mod_form = NULL,
  tp_col = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dataset_y}{data.table Filtered butterfly counts for species x over year y over all sites.}

\item{NbrSample}{integer Inherited from \link{flight_curve}, default=100.}

\item{GamFamily}{string Inherited from \link{flight_curve}, default='poisson', but can be 'nb' or 'quasipoisson'.}

\item{MaxTrial}{integer Inherited from \link{flight_curve}, default=3.}

\item{SpeedGam}{logical to use the \link[mgcv]{bam} method instead of the \link[mgcv]{gam} method.}

\item{OptiGam}{logical Set if the \link[mgcv]{bam} method should be used, default instead of the default \link[mgcv]{gam} method.}

\item{TimeUnit}{character The time-step for which the spline should be computed, 'd' day or 'w' week.}

\item{MultiVisit}{string Function to apply for summarising multiple counts within a time unit, 'max' or 'mean' (default).}

\item{weekday}{Integer for selected day of the week for weekly summary, default is 3 (Wednesday). [1-7] where 1 = Monday.}

\item{mod_form}{string with formula to be passed to the gam model, default null.}

\item{tp_col}{string or vector of string with additional variable used in the gam model, default null.}

\item{verbose}{a logical indicating if some “progress report” should be given.}

\item{...}{Additional parameters passed to gam or bam function from the \link[mgcv]{gam} package.}
}
\value{
A list with three objects, i) **f_curve**: a data.table with the flight curve \code{f_curve} with expected relative abundance, normalize to sum to one over a full season,
        ii) **f_model**: the resulting gam model \code{f_model} fitted on the count data and iii) **f_data**: a data.table with the data used to fit the GAM model. This is provide for one year 'y'.
}
\description{
fit_gam
fit a Generalized Additive Model to butterfly count data along a temporal variable and accounting for site effect when multiple are available.
}
\seealso{
\link{flight_curve}, \link[mgcv]{gam}, \link[mgcv]{bam}
}
\author{
Reto Schmucki - \email{retoshm@ceh.ac.uk}
}
\keyword{gam}
