% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{flight_curve}
\alias{flight_curve}
\title{flight_curve
Compute the annual flight curve from butterfly count data collated across multiple sites.}
\usage{
flight_curve(
  ts_season_count,
  NbrSample = 100,
  MinVisit = 3,
  MinOccur = 2,
  MinNbrSite = 1,
  MaxTrial = 3,
  GamFamily = "poisson",
  CompltSeason = TRUE,
  SelectYear = NULL,
  SpeedGam = TRUE,
  OptiGam = TRUE,
  KeepModel = TRUE,
  KeepModelData = TRUE,
  ConLikelihood = TRUE,
  TimeUnit = "d",
  MultiVisit = "mean",
  ...
)
}
\arguments{
\item{ts_season_count}{data.table with complete time series of count and season information returned by \link{ts_monit_count_site}}

\item{NbrSample}{integer setting the maximum number of site to use to compute the flight curve, default=100.}

\item{MinVisit}{integer setting the minimum number of visit required for a site to included in the computation, default=3.}

\item{MinOccur}{integer setting the minimum number of positive records (e.g. >= 1) observed over the year in a site default=2.}

\item{MinNbrSite}{integer setting the minimum number of site required to compute the flight curve, default=1.}

\item{MaxTrial}{integer setting the maximum number of trial to reach convergence of the model, default=3.}

\item{GamFamily}{string setting the distribution of the error term in the GAM, default='poisson', but can be 'nb' or 'quasipoisson'.}

\item{CompltSeason}{Logical to restrict computation of flight curve for years where the complete season has been sampled, default=TRUE.}

\item{SelectYear}{integer to select a specific year to compute the flight curve, default=NULL.}

\item{SpeedGam}{Logical to use the \link[mgcv]{bam} method instead of the \link[mgcv]{gam} method.}

\item{OptiGam}{Logical to set use bam when data are larger than 100 and gam for smaller dataset.}

\item{KeepModel}{Logical to keep model output in a list object named \code{flight_curve_model}.}

\item{KeepModelData}{Logical to keep the data used for the GAM.}

\item{ConLikelihood}{Logical to use the concentrated likelihood approach without site parameters (default = TRUE).}

\item{TimeUnit}{Character to define days 'd' or week 'w' as variable for the GAM.}

\item{MultiVisit}{function to be applied on multiple count within a single time unit, 'max' or 'mean' (default).}

\item{...}{additional parameters passed to gam or bam function from the \link[mgcv]{gam} package.}
}
\value{
A list with three objects, i) **pheno**: a vector with annual flight curves \code{f_pheno} with expected relative abundance, normalize to sum to one over a full season,
        ii) **model**: a list of the resulting gam models \code{f_model} fitted on the count data for each year and iii) **data**: a data.table with the data used to fit the GAM model.
}
\description{
flight_curve
Compute the annual flight curve from butterfly count data collated across multiple sites.
}
\seealso{
\link{fit_gam}

\link[mgcv]{gam}

\link[mgcv]{bam}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{curve}
\keyword{flight}
\keyword{gam,}
