% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{flight_curve}
\alias{flight_curve}
\title{flight_curve
Compute the annual flight curve from butterfly count data collated across multiple sites.}
\usage{
flight_curve(ts_season_count, NbrSample = 100, MinVisit = 3, MinOccur = 2,
  MinNbrSite = 1, MaxTrial = 3, FcMethod = "regionalGAM",
  GamFamily = "poisson", CompltSeason = TRUE, SelectYear = NULL,
  SpeedGam = TRUE, OptiGam = TRUE, KeepModel = TRUE, ...)
}
\arguments{
\item{ts_season_count}{data.table with complete time series of count and season information returned by \code{\link{ts_monit_count_site}}}

\item{NbrSample}{integer setting the maximum number of site to use to compute the flight curve, default=100.}

\item{MinVisit}{integer setting the minimum number of visit required for a site to included in the computation, default=3.}

\item{MinOccur}{integer setting the minimum number of positive records (e.g. >= 1) observed over the year in a site default=2.}

\item{MinNbrSite}{integer setting the minimum number of site required to compute the flight curve, default=1.}

\item{MaxTrial}{integer setting the maximum number of trial to reach convergence of the model, default=3.}

\item{FcMethod}{string defining the method to be used for computation of the flight curve, default='regionalGAM' (no other method is available yet).}

\item{GamFamily}{string setting the distribution of the error term in the GAM, default='poisson', but can be 'nb' or 'quasipoisson'.}

\item{CompltSeason}{Logical to restrict computation of flight curve for years where the complete season has been sampled, default=TRUE.}

\item{SelectYear}{integer to select a specific year to compute the flight curve, default=NULL.}

\item{SpeedGam}{Logical to use the \code{link[mgcv]{bam}} method instead of the \code{\link[mgcv]{gam}} method.}

\item{OptiGam}{Logical to set use bam when data are larger than 100 and gam for smaller dataset}

\item{KeepModel}{Logical to keep model output in a list object named \code{flight_curve_model}}

\item{...}{additional parameters passed to gam or bam function from the \code{\link[mgcv]{gam}} package.}
}
\value{
A list with two object, i) a vector with annual fligth curves \code{f_pheno} with expected relative abudance, normalize to sum to one over a full season,
        and ii) a list of the resulting gam models \code{f_model} fitted on the count data for each year.
}
\description{
flight_curve
Compute the annual flight curve from butterfly count data collated across multiple sites.
}
\seealso{
\code{\link{fit_gam}}

\code{\link[mgcv]{gam}}

\code{\link[mgcv]{bam}}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{curve}
\keyword{flight}
\keyword{gam,}
