% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{get_nny}
\alias{get_nny}
\title{get_nny
find nearest year with flight period}
\usage{
get_nny(x, y)
}
\arguments{
\item{x}{year to find nearest flight period.}

\item{y}{years with availabe flight period.}
}
\value{
z a value of the nearest year
}
\description{
get_nny
find nearest year with flight period
}
\seealso{
\link{check_pheno}
}
\author{
Reto Schmucki - \email{reto.schmucki@mail.mcgill.ca}
}
\keyword{curve}
\keyword{flight}
