% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_index_modelling.r
\name{impute_count}
\alias{impute_count}
\title{impute_count}
\usage{
impute_count(
  ts_season_count,
  ts_flight_curve,
  TimeUnit = "d",
  MultiVisit = "mean",
  weekday = 3,
  sp = NULL,
  YearLimit = NULL,
  SelectYear = NULL,
  CompltSeason = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{ts_season_count}{data.table Time-series of counts for a specific species across all sites as returned by \link{ts_monit_count_site}.}

\item{ts_flight_curve}{pheno_curve class object or data.table with light curves and relative abundances (NM) for a specific species as returned by \link{flight_curve}.}

\item{TimeUnit}{character The time-step for which the spline should be computed, 'd' day or 'w' week.}

\item{MultiVisit}{string Function to apply for summarising multiple counts within a time unit, 'max' or 'mean' (default).}

\item{weekday}{Integer for selected day of the week for weekly summary, default is 3 (Wednesday). [1-7] where 1 = Monday.}

\item{sp}{integer or string Species ID or name.}

\item{YearLimit}{integer Define the span of years (+/- number of year) to look for a flight curve, if NULL no restriction is set.}

\item{SelectYear}{integer Select a specific year to compute the flight curve, default=NULL.}

\item{CompltSeason}{logical Restrict computation of flight curve for years where the complete season was sampled, default=TRUE.}

\item{verbose}{a logical indicating if some “progress report” should be given.}
}
\value{
A data.table based on the entry count data, augmented with site indices 'SINDEX' and imputed weekly count 'IMPUTED_COUNT'.
}
\description{
impute_count
}
\details{
Site indices can be extracted from the data.table returned by this function. The site index is currently computed by adjusting the count by the proportion of the flight curve covered by the visits.
}
\seealso{
\link{flight_curve}
}
\author{
Reto Schmucki - \email{retoshm@ceh.ac.uk}
}
\keyword{curve}
\keyword{flight}
\keyword{index,}
\keyword{site}
