% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbms_toolbox.r
\name{ts_date_seq}
\alias{ts_date_seq}
\title{ts_date_seq
Generate a time-series with dates starting from January of the starting year to December of the ending years.}
\usage{
ts_date_seq(InitYear = 1970, LastYear = format(Sys.Date(), "\%Y"))
}
\arguments{
\item{InitYear}{First year of the time-series, four digits format (e.g. 1987).}

\item{LastYear}{Last year of the time-series, if not provided, the current year is used.}
}
\value{
Returns a POSIXct vector with the format 'YYYY-MM-DD HH:MM:SS'
}
\description{
ts_date_seq
Generate a time-series with dates starting from January of the starting year to December of the ending years.
}
\author{
Reto Schmucki - \email{retoshm@ceh.ac.uk}
}
\keyword{time-series}
